Separate doclist related files
The doclist is in one file, because it involves a lock.
