+++ oviewer/utils.go
@@ -1,5 +1,6 @@
 package oviewer
 
+// max returns the larger value of the argument.
 func max(a, b int) int {
 	if a > b {
 		return a
@@ -7,6 +8,7 @@ func max(a, b int) int {
 	return b
 }
 
+// max returns the smaller value of the argument.
 func min(a, b int) int {
 	if a < b {
 		return a
@@ -14,6 +16,7 @@ func min(a, b int) int {
 	return b
 }
 
+// removeStr removes the value of the specified string from slice.
 func removeStr(list []string, s string) []string {
 	if len(s) == 0 {
 		return list
@@ -27,6 +30,7 @@ func removeStr(list []string, s string) []string {
 	return list
 }
 
+// removeStr removes the specified int value from slice.
 func removeInt(list []int, c int) []int {
 	for n, l := range list {
 		if l == c {
@@ -36,6 +40,7 @@ func removeInt(list []int, c int) []int {
 	return list
 }
 
+// containsInt returns true if the specified int is included.
 func containsInt(list []int, e int) bool {
 	for _, n := range list {
 		if e == n {
